pipeline/codeql/requirements.txt
python-sonarqube-api==2.0.5
requests==2.31.0
GitPython==3.1.40
boto3==1.35.32
kafka-python==2.0.2
